+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v0.7
+- Replaced SHA256 with Argon2 as encryption algorithm for better protection of tokens
+    - NOTE: this is a breaking change. Tokens imported with previous app versions will not work and will need to be reimported.
+- Added a security step with PIN / Password entering requirement to disable app protection
+
 # v0.6
 - Added open on startup feature
 - New workflow for app updates, including in app changelog for the new release and an automated portable downloader in the User download folder